CRUD FE bugs fixed. Devices handling updated v0.2.

+++ Thss0.Devices/Program.cs
@@ -1,6 +1,6 @@
 ﻿using Thss0.Devices;
 
-const ushort clientsQty = 6;
+const int clientsQty = 6;
 var clients = new Device[clientsQty];
 for (int i = 0; i < clientsQty; i++)
 {
